backend-google-ai-api/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
flask-cors==4.0.0
google-cloud-aiplatform==1.26.1
google-cloud-bigquery==3.10.0
flask-jwt-extended==4.4.4
PyYAML==6.0
google-auth==2.17.3
python-dateutil==2.8.2
pydantic==1.10.7